+++ src/api/RolesApi.ts
@@ -4,6 +4,7 @@ import { useQuery } from "@tanstack/react-query";
 
 export interface IRoles {
   isFocalPoint: boolean;
+  isAdmin: boolean;
 }
 
 const transformGroups = (data: ISiteGroupInfo[]): IRoles => {
@@ -13,8 +14,13 @@ const transformGroups = (data: ISiteGroupInfo[]): IRoles => {
     ? true
     : false;
 
+  const isAdmin = data.find((value) => value.Title.endsWith("Owners"))
+    ? true
+    : false;
+
   const roles = {
     isFocalPoint: isFocalPoint,
+    isAdmin: isAdmin,
   };
 
   return roles;
@@ -27,6 +33,20 @@ const getRoles = () => {
     const groups: ISiteGroupInfo[] = [
       {
         Title: "CAFTOP Narrative Focal Points",
+        AllowMembersEditMembership: false,
+        AllowRequestToJoinLeave: false,
+        AutoAcceptRequestToJoinLeave: false,
+        Description: "",
+        Id: 0,
+        IsHiddenInUI: false,
+        LoginName: "",
+        OnlyAllowMembersViewMembership: false,
+        OwnerTitle: "",
+        PrincipalType: 0,
+        RequestToJoinLeaveEmailSetting: null,
+      },
+      {
+        Title: "CAFTOP Narrative Owners",
         AllowMembersEditMembership: false,
         AllowRequestToJoinLeave: false,
         AutoAcceptRequestToJoinLeave: false,