+++ CHANGELOG.md
@@ -28,6 +28,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Identifier added to left navigation panel to let user know what CAFTOP they are currently viewing/editing
+- Added mew role of Admin, which are those users in the Owners group. Allowed them to do any action a Focal Point could
+- Added ability for Admin to Delete CAFTOPs
 
 ## [1.0.3] - 2025-02-13